// src/lib/api/generated/models/category.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { CategoryStatus } from './categoryStatus';

export interface Category {
  auction_items_count?: number;
  created_at: string;
  /** @nullable */
  description?: string | null;
  id: number;
  /** @nullable */
  image?: string | null;
  name: string;
  /** @nullable */
  parent_id?: number | null;
  slug: string;
  sort_order: number;
  status: CategoryStatus;
  updated_at: string;
}