// src/lib/api/generated/models/getAdminCategoriesParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { GetAdminCategoriesStatus } from './getAdminCategoriesStatus';

export type GetAdminCategoriesParams = {
page?: number;
per_page?: number;
search?: string;
status?: GetAdminCategoriesStatus;
parent_id?: number;
};