// src/lib/api/generated/models/updateCategoryRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { UpdateCategoryRequestStatus } from './updateCategoryRequestStatus';

export interface UpdateCategoryRequest {
  description?: string;
  image?: string;
  name?: string;
  parent_id?: number;
  sort_order?: number;
  status?: UpdateCategoryRequestStatus;
}