// src/lib/api/generated/models/getAdminCategoriesStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */

export type GetAdminCategoriesStatus = typeof GetAdminCategoriesStatus[keyof typeof GetAdminCategoriesStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetAdminCategoriesStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;